Add Configuration factory implementation.
