+++ src/main/kotlin/config/Configuration.kt
@@ -1,11 +1,5 @@
 package config
 
-class Configuration : ConfigurationProvider() {
-    fun createObjectFromJSON(): ConfigurationObject {
-        return ConfigurationProvider().getObjectFromJSONFile()
-    }
-
-    fun createObjectFromYAML(): ConfigurationObject {
-        return ConfigurationProvider().getObjectFromYAMLFile()
-    }
+interface Configuration {
+  fun getConfig(): ConfigurationObject
 }
