+++ src/main/kotlin/config/ConfigurationProvider.kt
@@ -1,25 +1,16 @@
 package config
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
-import com.fasterxml.jackson.module.kotlin.KotlinModule
-import com.fasterxml.jackson.module.kotlin.readValue
-
 open class ConfigurationProvider {
-    private val JSONpath = "config/json/base_configuration"
-    private val YAMLpath = "config/yaml/base_configuration"
-    private val objectMapperJSON = ObjectMapper().registerModule(KotlinModule())
-    private val objectMapperYAML = ObjectMapper(YAMLFactory()).registerModule(KotlinModule())
+  enum class ConfigSource {
+    YAML, JSON
+  }
 
-    fun getObjectFromJSONFile(): ConfigurationObject {
-        return Thread.currentThread().contextClassLoader.getResourceAsStream(JSONpath).use {
-            objectMapperJSON.readValue(it)
-        }
+  companion object {
+    fun setConfigType(type: ConfigSource): Configuration {
+      return when (type) {
+        ConfigSource.YAML -> YAMLConfiguration()
+        ConfigSource.JSON -> JSONConfiguration()
+      }
     }
-
-    fun getObjectFromYAMLFile(): ConfigurationObject {
-        return Thread.currentThread().contextClassLoader.getResourceAsStream(YAMLpath).use {
-            objectMapperYAML.readValue(it)
-        }
-    }
+  }
 }
